dmaengine: idxd: idxd_desc.id is now a u16
This is going to be packed into the cookie. It does not need to be
negative or larger than u16.

+++ drivers/dma/idxd/idxd.h
@@ -328,7 +328,7 @@ struct idxd_desc {
 	struct dma_async_tx_descriptor txd;
 	struct llist_node llnode;
 	struct list_head list;
-	int id;
+	u16 id;
 	int cpu;
 	struct idxd_wq *wq;
 };
